7_docker_jenkins/interface_python/requirements.txt
allure-pytest
appium-python-client
pytest-testconfig
requsts
selenium
urllib3